Elm pages3 (#89)
* fix devcontainer

* Port to elm-pages V3

* Move Post to src

* chore: Update site.webmanifest and favicon images

This commit updates the site.webmanifest file and the favicon images. The site.webmanifest file now includes the correct URLs for the icons, and the favicon images have been added in the sizes of 16x16 and 32x32 pixels. These changes ensure that the website has proper icons and displays correctly on different devices.

* Add resonsiness :D